+++ changelog.md
@@ -6,6 +6,7 @@
 1. Added confirmation for buying packs with UCP
 1. Added a setting to switch minigame controls to WASD (disabled by default)
 1. Added a random avatar button
+1. Added setting to prefer shiny on import (enabled by default)
 ### Fixes
 1. Fixed bug with base card skin setting
 ### Plugins